+++ gammapy/data/observations.py
@@ -6,7 +6,6 @@ import logging
 import warnings
 from itertools import zip_longest
 import numpy as np
-import scipy.cluster.hierarchy as sch
 import astropy.units as u
 from astropy.coordinates import SkyCoord
 from astropy.io import fits
@@ -750,60 +749,26 @@ class Observations(collections.abc.MutableSequence):
                     features[ko, kf] = psf_radius.value
         return features
 
-    def hierarchical_clustering(
-        self, features, linkage_kwargs=None, fcluster_kwargs=None, standard_scaler=False
-    ):
-        """Hierarchical clustering of obsevrations using given features.
+    def split(self, groups_id):
+        """Split obsevations in multiple groups of observations
 
         Parameters
         ----------
-        features : array
-            (N x M) array with N observations and M features.
-        linkage_kwargs : dict
-            Arguments forwarded to `scipy.cluster.hierarchy.linkage`
-        fcluster_kwargs : dict
-            Arguments forwarded to `scipy.cluster.hierarchy.fcluster`
-        standard_scaler : bool
-            Standardize features by removing the mean and scaling to unit variance.
-            Default is False.
+        cluster_id : array
+            Array of cluster ID
 
         Returns
         -------
         obs_clusters : list of `~gammapy.data.Observations`
-            List of Observations instance, one instance for each cluster.
-        ind_clusters : array
-            Array of cluster ID
-        features : array
-            Scaled features
+            List of Observations instance, one instance for each group.
         """
-
-        if standard_scaler:
-            features = features.copy()
-            for kf in range(features.shape[1]):
-                features[:, kf] = (features[:, kf] - features[:, kf].mean()) / features[
-                    :, kf
-                ].std()
-
-        default_linkage_kwargs = dict(method="ward", metric="euclidean")
-        if linkage_kwargs is not None:
-            default_linkage_kwargs.update(linkage_kwargs)
-
-        pairwise_distances = sch.distance.pdist(features)
-        linkage = sch.linkage(pairwise_distances, **default_linkage_kwargs)
-
-        default_fcluster_kwargs = dict(criterion="maxclust", t=3)
-        if fcluster_kwargs is not None:
-            default_fcluster_kwargs.update(fcluster_kwargs)
-        ind_clusters = sch.fcluster(linkage, **default_fcluster_kwargs)
-
         obs_clusters = []
-        for ind in np.unique(ind_clusters):
+        for ind in np.unique(groups_id):
             observations = self.__class__(
-                [obs for k, obs in enumerate(self) if ind_clusters[k] == ind]
+                [obs for k, obs in enumerate(self) if groups_id[k] == ind]
             )
             obs_clusters.append(observations)
-
-        return obs_clusters, ind_clusters, features
+        return obs_clusters
 
 
 class ObservationChecker(Checker):
